TODO-PHASE02/backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlmodel==0.0.22
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-dotenv==1.0.1

# HTTP Client
httpx==0.28.1

# Email Validation
pydantic[email]==2.10.4

# Utilities
python-multipart==0.0.19
